more README, move LOCATION_CHANGE and SERVER_RENDER to router package
