+++ src/router/index.js
@@ -4,3 +4,8 @@ export { default as NotFound } from './NotFound';
 export { default as Redirect } from './Redirect';
 export { default as RouteLayout } from './RouteLayout';
 export { default as SyncContainer } from './SyncContainer';
+export {
+	LOCATION_CHANGE,
+	SERVER_RENDER,
+	locationChange,
+} from './routeActionCreators';
